Factored lobe rendering into separate class.
